+++ meta_package_manager/managers/composer.py
@@ -159,6 +159,16 @@ class Composer(PackageManager):
 
         return matches
 
+    def install(self, package_id):
+        """Install one package.
+
+        .. code-block:: shell-session
+
+            ► composer global install illuminate/contracts
+
+        """
+        return self.run_cli(self.global_args, "install", package_id)
+
     @property
     def outdated(self):
         """Fetch outdated packages from ``composer global outdated`` output.